Added new light system
Put a NBLight component on every object that needs to cast light. The gradient
start/end are in unity length units. They support enabling and disabling on
the fly.
To actually see any effect of it, put a LightOverlay on the main camera.
There is currently a hardcoded limit of 20 lights, just because. Probably easy
to increase.